rename converter to match its filetype
